Use a single convention for variable shorthands.
For indexes or elements from an iteration, "i" is used, "j" is used, where "i"
is taken, "k" is used where both "i" and "j" are taken. For callbacks "f" is
used, "g" and "h" may be used where "f" is taken.  For DOMEvents "e" is used.
For objects, that configure the behavior of a function "o" is used.  For wrapper
functions "a", "b", "c" and "d" are used. For functions, that pass parameters to
a callback, "x", "y" and "z" are used. For values that will be returned, "r" is
used. For natural numbers passed to a function, "n" and "m" are used. For
functions that get a list, "l" is used, "p" and "q" are pointers into the list.
For the blindingly obvious, naming which would be redundant redundancy "_" is
used.  The scope of all vars with less than three letters in the name shall fit
on a reasonably sized screen.